+++ sdk/loadtesting/playwright/src/utils/utils.ts
@@ -165,32 +165,6 @@ export const fetchOrValidateAccessToken = async (credential?: TokenCredential): 
   return getAccessToken()!;
 };
 
-const getPackageVersionFromFolder = (folder: string): string => {
-  try {
-    // eslint-disable-next-line @typescript-eslint/no-require-imports
-    const version = require(path.join(__dirname, folder, "package.json")).version;
-    return version;
-  } catch (error) {
-    coreLogger.error("Error fetching package version:", error);
-    return "";
-  }
-};
-
-export const getPackageVersion = (): string => {
-  // hacky way to get package version
-  // try from dist folder first (customer perspective)
-  const distVersion = getPackageVersionFromFolder("../../../");
-  if (distVersion) {
-    return distVersion;
-  }
-  // if not found, try from src folder (internal test suite)
-  const srcVersion = getPackageVersionFromFolder("../../");
-  if (srcVersion) {
-    return srcVersion;
-  }
-  return "unknown-version";
-};
-
 export const getVersionInfo = (version: string): VersionInfo => {
   const regex = /^(\d+)(?:\.(\d+))?(?:\.(\d+))?/;
   const match = version.match(regex);
